Create LoginfoWrite.js
That would be the last of the database files that need to be filled. These files should be filled sometime tomorrow. In case there's any questions, I wrote this code on repl.it, only for it to do some dumb stuff and make me unable to view the preview. I will likely continue this on another place similar and copy the code here, as it seems it doesn't have a preview here either.